+++ driver/src/main/scala/com/stratio/sparta/driver/stage/OutputStage.scala
@@ -17,17 +17,17 @@ package com.stratio.sparta.driver.stage
 
 import java.io.Serializable
 
-import com.stratio.sparta.driver.utils.ReflectionUtils
 import com.stratio.sparta.sdk.pipeline.output.Output
 import com.stratio.sparta.sdk.pipeline.schema.SpartaSchema
 import com.stratio.sparta.serving.core.constants.AppConstant
 import com.stratio.sparta.serving.core.models.policy.{PhaseEnum, PolicyElementModel}
+import com.stratio.sparta.serving.core.utils.ReflectionUtils
 
 trait OutputStage extends BaseStage {
   this: ErrorPersistor =>
 
-  def outputStage(schemas: Seq[SpartaSchema],
-                  refUtils: ReflectionUtils): Seq[Output] = policy.outputs.map(o => {
+  def outputStage(schemas: Seq[SpartaSchema], refUtils: ReflectionUtils): Seq[Output] =
+    policy.outputs.map(o => {
     val schemasAssociated = schemas.filter(tableSchema => tableSchema.outputs.contains(o.name))
     createOutput(o, schemasAssociated, refUtils)
   })