+++ serving-api/src/main/scala/com/stratio/sparta/serving/api/utils/ClusterSparkFilesUtils.scala
@@ -21,11 +21,11 @@ import com.stratio.sparta.serving.core.config.SpartaConfig
 import com.stratio.sparta.serving.core.constants.AppConstant
 import com.stratio.sparta.serving.core.helpers.JarsHelper
 import com.stratio.sparta.serving.core.models.policy.PolicyModel
-import com.stratio.sparta.serving.core.utils.{CheckpointUtils, HdfsUtils, PolicyUtils}
+import com.stratio.sparta.serving.core.utils.{CheckpointUtils, HdfsUtils}
 
 import scala.util.Try
 
-case class ClusterSparkFilesUtils(policy: PolicyModel, hdfs: HdfsUtils) extends PolicyUtils with CheckpointUtils {
+case class ClusterSparkFilesUtils(policy: PolicyModel, hdfs: HdfsUtils) extends CheckpointUtils {
 
   private val hdfsConfig = SpartaConfig.getHdfsConfig.get
   private val host = hdfsConfig.getString(AppConstant.HdfsMaster)