+++ serving-core/src/test/scala/com/stratio/sparta/serving/core/utils/PolicyUtilsTest.scala
@@ -16,49 +16,17 @@
 
 package com.stratio.sparta.serving.core.utils
 
-import com.stratio.sparta.serving.core.models.policy.{OutputFieldsModel, PolicyModel, UserJar}
+import com.stratio.sparta.serving.core.models.policy.PolicyModel
 import org.junit.runner.RunWith
 import org.mockito.Mockito._
 import org.scalatest.junit.JUnitRunner
 
 @RunWith(classOf[JUnitRunner])
-class PolicyUtilsTest extends BaseUtilsTest
-  with PolicyUtils {
+class PolicyUtilsTest extends BaseUtilsTest with PolicyUtils {
 
   val utils = spy(this)
   val basePath = "/samplePath"
   val aggModel: PolicyModel = mock[PolicyModel]
-
-  "PolicyUtils" should {
-    "return files" in {
-      when(aggModel.userPluginsJars).thenReturn(Seq(UserJar("path1"), UserJar("path2")))
-
-      val files = jarsFromPolicy(aggModel)
-
-      files.map(_.getName) shouldBe(Seq("path1", "path2"))
-    }
-
-    "return empty Seq" in {
-      when(aggModel.userPluginsJars).thenReturn(Seq(UserJar("")))
-
-      val files = jarsFromPolicy(aggModel)
-
-      files.size shouldBe(0)
-    }
-  }
-
-  "PolicyUtils.savePolicyInZk" should {
-
-    "write policy when this does not exist" in {
-      doReturn(getPolicyModel())
-        .when(utils)
-        .writePolicy(getPolicyModel(), curatorFramework)
-
-      utils.savePolicyInZk(policy = getPolicyModel(), curatorFramework)
-
-      verify(utils).writePolicy(getPolicyModel(), curatorFramework)
-    }
-  }
 
   "PolicyUtils.policyWithId" should {
     "return a policy with random UUID when there is no set id yet" in {
@@ -89,8 +57,8 @@ class PolicyUtilsTest extends BaseUtilsTest
         getPolicyModel(id = Some("id#2")),
         getPolicyModel(id = Some("id#3"))))
         .when(utils)
-        .getPolicies(curatorFramework, withFragments = false)
-      utils.existsByNameId(name = "myName", id = Some("existingID"), curatorFramework).get should be(
+        .getPolicies(withFragments = false)
+      utils.existsPolicyByNameId(name = "myName", id = Some("existingID")).get should be(
         getPolicyModel(id = Some("existingID")))
     }
 
@@ -103,10 +71,9 @@ class PolicyUtilsTest extends BaseUtilsTest
         getPolicyModel(id = Some("id#2")),
         getPolicyModel(id = Some("id#3"))))
         .when(utils)
-        .getPolicies(curatorFramework, withFragments = false)
+        .getPolicies(withFragments = false)
 
-      val actualPolicy: PolicyModel = utils.existsByNameId(name = "MYNAME", id = None,
-        curatorFramework).get
+      val actualPolicy: PolicyModel = utils.existsPolicyByNameId(name = "MYNAME", id = None).get
 
       actualPolicy.name should be("myname")
       actualPolicy.id.get should be("id#1")
@@ -121,9 +88,9 @@ class PolicyUtilsTest extends BaseUtilsTest
         getPolicyModel(id = Some("id#2")),
         getPolicyModel(id = Some("id#3"))))
         .when(utils)
-        .getPolicies(curatorFramework, withFragments = true)
+        .getPolicies(withFragments = true)
 
-      utils.existsByNameId(name = "noName", id = None, curatorFramework) should be(None)
+      utils.existsPolicyByNameId(name = "noName", id = None) should be(None)
     }
 
     "return none when there is some error or exception" in {
@@ -132,9 +99,9 @@ class PolicyUtilsTest extends BaseUtilsTest
         .existsPath
       doThrow(new RuntimeException)
         .when(utils)
-        .getPolicies(curatorFramework, withFragments = true)
+        .getPolicies(withFragments = true)
 
-      utils.existsByNameId(name = "noName", id = None, curatorFramework) should be(None)
+      utils.existsPolicyByNameId(name = "noName", id = None) should be(None)
     }
 
     "return none when path not does not exists" in {
@@ -142,7 +109,7 @@ class PolicyUtilsTest extends BaseUtilsTest
         .when(utils)
         .existsPath
 
-      utils.existsByNameId(name = "noName", id = None, curatorFramework) should be(None)
+      utils.existsPolicyByNameId(name = "noName", id = None) should be(None)
     }
   }
 }