+++ web/src/scripts/controllers/inputs.js
@@ -20,9 +20,9 @@
     .module('webApp')
     .controller('InputsCtrl', InputsCtrl);
 
-  InputsCtrl.$inject = ['FragmentFactory', '$filter', '$uibModal', 'UtilsService', 'TemplateFactory', 'PolicyFactory'];
+  InputsCtrl.$inject = ['FragmentFactory', '$filter', '$uibModal', 'UtilsService', 'TemplateFactory', 'FragmentService'];
 
-  function InputsCtrl(FragmentFactory, $filter, $uibModal, UtilsService, TemplateFactory, PolicyFactory) {
+  function InputsCtrl(FragmentFactory, $filter, $uibModal, UtilsService, TemplateFactory, FragmentService) {
     /*jshint validthis: true*/
     var vm = this;
 
@@ -32,6 +32,7 @@
     vm.editInput = editInput;
     vm.duplicateInput = duplicateInput;
     vm.deleteErrorMessage = deleteErrorMessage;
+    vm.getFragmentIcon = FragmentService.getFragmentIcon;
     vm.inputsData = undefined;
     vm.inputTypes = [];
     vm.errorMessage =  {type: 'error',text: '', internalTrace: ''};
@@ -155,18 +156,18 @@
       }
     }
 
-    function createInputModal(newInputTemplateData) {
+    function createInputModal(creationFragmentData) {
       var modalInstance = $uibModal.open({
         animation: true,
         templateUrl: 'templates/fragments/fragment-details.tpl.html',
         controller: 'NewFragmentModalCtrl as vm',
         size: 'lg',
         resolve: {
-          item: function () {
-            return newInputTemplateData;
+          creationFragmentData: function () {
+            return creationFragmentData;
           },
-          fragmentTemplates: function () {
-            return TemplateFactory.getNewFragmentTemplate(newInputTemplateData.fragmentType);
+          fragmentTemplate: function () {
+            return TemplateFactory.getNewFragmentTemplate(creationFragmentData.fragmentType);
           }
         }
       });
@@ -184,10 +185,10 @@
         controller: 'EditFragmentModalCtrl as vm',
         size: 'lg',
         resolve: {
-          item: function () {
+          creationFragmentData: function () {
             return editInputData;
           },
-          fragmentTemplates: function () {
+          fragmentTemplate: function () {
             return TemplateFactory.getNewFragmentTemplate(editInputData.fragmentSelected.fragmentType);
           }
         }
@@ -213,7 +214,7 @@
         controller: 'DeleteFragmentModalCtrl as vm',
         size: 'lg',
         resolve: {
-          item: function () {
+          fragmentTemplate: function () {
             return input;
           }
         }
@@ -236,7 +237,7 @@
         controller: 'DuplicateFragmentModalCtrl as vm',
         size: 'lg',
         resolve: {
-          item: function () {
+          fragmentTemplate: function () {
             return InputData;
           }
         }